+++ CHANGELOG.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
 - Translate duration strings to a language and format that makes sense to the user (for private responses) or to the guild (for public responses).
 - `/quo stats` no longer repsonds publicly, even when the channel of invocation is the queue channel.
+- `/now-playing` has been renamed to `/nowplaying`. The command retains `now-playing` as an alias, in case users wish to do things the hard way.
 
 ## [1.7.1] - 2022-08-14
 ### Added